requirements.txt
# Telegram Bot
aiogram==3.13.1

# Загрузка данных
yfinance==0.2.51
curl_cffi>=0.5.0

# ML модели
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3

# Статистические модели
statsmodels==0.14.4
prophet==1.1.6

# Нейронные сети (PyTorch для Python 3.13)
torch>=2.6.0
torchvision>=0.21.0

# Визуализация
matplotlib==3.9.2

# Утилиты
scipy==1.14.1
python-dotenv==1.0.1
cachetools==5.5.0
